Don't call class_create (and other functions) with a local pointer.
The ndp_name will get be invalid as soon as vbc_turning_ndp_init exits, but since class_create and other functions store a pointer to the name argument this will cause crashes.

So instead of pointing to a local variable, point to a global string instead.

Profile_size check is there just in case.

Signed-off-by: Matti Kosola <matti.kosola@seafarix.com>